Add sidebar for navigations
This commit adds a left nav sidebar and move links from the header.
The change includes adjusting the left paddings of the pages to account
for the size of the sidebar.
